+++ loyaltyprogrambackend/src/main/java/it/cs/unicam/dicygroup/loyaltyprogrambackend/piano/regola/RegolaPunti.java
@@ -8,7 +8,7 @@ public class RegolaPunti implements Regola {
     }
 
     @Override
-    public Double applica(Double importo) {
-        return importo * moltiplicatore;
+    public Integer applica(Double importo) {
+        return (int) (importo * moltiplicatore);
     }
 }
